// src/api-client/models/api-response-page-response-product-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PageResponseProductResponse } from './page-response-product-response';

/**
 * 
 * @export
 * @interface ApiResponsePageResponseProductResponse
 */
export interface ApiResponsePageResponseProductResponse {
    /**
     * 
     * @type {number}
     * @memberof ApiResponsePageResponseProductResponse
     */
    'code'?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiResponsePageResponseProductResponse
     */
    'message'?: string;
    /**
     * 
     * @type {PageResponseProductResponse}
     * @memberof ApiResponsePageResponseProductResponse
     */
    'result'?: PageResponseProductResponse;
}